Added Module file and refactored some code on Student, Apprentice Teacher and Senior Teacher for redundancy.
