+++ student.rb
@@ -1,16 +1,19 @@
-class Student
+require_relative "modules"
+
+
+class Student 
   attr_reader :age, :phase
   attr_accessor :name
+  include Offer_high_five
 
   def initialize(options = {})
+  
     @phase = 1
+
     @age = options.fetch(:age, 0)
     @name = options.fetch(:name, "")
   end
 
-  def offer_high_five
-    "High five!"
-  end
 
   def set_phase(num)
     response = ""
